requirements.txt
opencv-python
pyyaml
protobuf
paho-mqtt
